// src/types/list.movies.dto.ts
interface IListMovie {
  id: string;
  certificate: string;
  createdAt: string;
  director: string;
  genre: string;
  gross: string;
  imdb_raiting: number;
  meta_score: number;
  no_of_votes: number;
  overview: string;
  poster_link: string;
  released_year: number;
  runtime: string;
  series_title: string;
  star1: string;
  star2: string;
  star3: string;
  star4: string;
  updatedAt: string;
}

export default IListMovie;